moved composer.ts
