+++ extension/chrome/elements/reply_pubkey_mismatch.ts
@@ -5,7 +5,7 @@
 import { Store, Subscription } from '../../js/common/platform/store.js';
 import { Att } from '../../js/common/core/att.js';
 import { Ui, Env } from '../../js/common/browser.js';
-import { Composer } from '../../js/common/composer.js';
+import { Composer } from '../../js/common/composer/composer.js';
 import { Api } from '../../js/common/api/api.js';
 import { BrowserMsg } from '../../js/common/extension.js';
 import { Catch } from '../../js/common/platform/catch.js';
